pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "portfolio"
version = "1.1.0"
description = "A PortFolio website build with the CMS Wagtail"
authors = ["Yannick Chabbert <yannick.chabbert.it@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.5"
wagtail = "^4.1.1"
django-ipware = "^4.0.2"
pyyaml = "^6.0"
django-compressor = "^4.1"
redis = "^4.3.4"
hiredis = "^2.0.0"

[tool.poetry.group.test.dependencies]
isort = "^5.10.1"
tblib = "^1.7.0"
tox = "^3.27.1"
coverage = {extras = ["toml"], version = "^6.5.0"}
pylama = {extras = ["toml"], version = "^8.4.1"}

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
django-debug-toolbar = "^3.7.0"
pre-commit = "^2.20.0"

[tool.pylama]
skip = "*/.tox/*,*/migrations/*,*/settings/*,*/static/*"
linters = "pycodestyle,pyflakes,mccabe"

[tool.pylama.linter.pycodestyle]
max_line_length = 119

[tool.pylama.linter.mccabe]
complexity = 15

[tool.isort]
combine_as_imports = true
multi_line_output = 5
include_trailing_comma = true
skip = ".tox,migrations,docs,media,static"
sections = "FUTURE,STDLIB,DJANGO,WAGTAIL,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
known_django = "django"
known_wagtail = "wagtail"
known_first_party = "portfolio"

[tool.coverage.run]
source = [ "portfolio" ]
branch = true
omit = [
    "manage.py",
    "portfolio/wsgi.py",
    "portfolio/settings/**",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise AssertionError",
    "raise NotImplementedError",
    "def __str__",
    "def __repr__",
]
